requirements.txt
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cryptography==45.0.3
cssselect2==0.8.0
decorator==4.4.2
defusedxml==0.7.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
docutils==0.16
ebaysdk==2.1.5
et_xmlfile==2.0.0
freezegun==0.3.15
gevent==24.2.1
greenlet==3.0.3
html5lib==1.1
idna==2.10
isodate==0.6.1
Jinja2==3.1.2
libsass==0.22.0
lxml==5.4.0
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
num2words==0.5.9
numpy==2.2.6
ofxparse==0.21
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
passlib==1.7.4
pillow==10.2.0
polib==1.1.0
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydot==1.4.2
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==5.5.0
PyPDF2==2.12.1
pyserial==3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-ldap==3.4.4
python-stdnum==1.16
pytz==2024.1
pyusb==1.2.1
PyYAML==6.0.2
qrcode==6.1
reportlab==4.1.0
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==70.3.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.0.7
utils==1.0.2
vobject==0.9.6.1
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.43.0
xhtml2pdf==0.2.17
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==4.0.0
zope.event==5.0
zope.interface==6.4.post2
